♻️ refactor: extract useDashboardFavorite hook from DashboardGrid #14
